Refactoring

+++ Strand/src/Strand/Renderer/Buffer.h
@@ -33,7 +33,7 @@ namespace Strand {
 		std::string Name;
 		ShaderDataType Type;
 		uint32_t Size;
-		uint32_t Offset;
+		size_t  Offset;
 		bool Normalized;
 		BufferElement(ShaderDataType type, const std::string& name, bool normalized = false)
 			: Name(name), Type(type), Size(ShaderDataTypeSize(type)), Offset(0), Normalized(normalized)
@@ -84,7 +84,8 @@ namespace Strand {
 	private:
 		void CalculateOffsetsAndStride()
 		{
-			uint32_t offset = 0;
+			//TODO
+			size_t  offset = 0;
 			m_Stride = 0;
 			for (auto& element : m_Elements)
 			{
